test(04-03): add Parquet-variant tests to test_schemas + test_aggregates (D-19, D-20)
- tests/test_schemas.py: appended module-scoped derived_dir fixture
  (rebuild once per module) + parametrised test_parquet_grain_schema
  over all five grains. Asserts column set + order matches the Pydantic
  model's field-declaration order (D-10), then validates every row via
  model.model_validate(). Formal TEST-02 (D-19).
- tests/test_aggregates.py: appended fixtures for derived_dir +
  station_month + annual_summary + by_technology + by_allocation_round
  and three row-conservation tests comparing sum(lower grain by year)
  against the matching year total in annual_summary via
  pd.testing.assert_series_equal (exact equality). Formal TEST-03 (D-20).
- Rule 1 fix: src/uk_subsidy_tracker/schemes/cfd/aggregation.py — cast
  `year` column to int64 in all three rollup builders to match the
  AnnualSummaryRow / ByTechnologyRow / ByAllocationRoundRow dtype
  declarations (int64 per StationMonthRow schema). Without the cast,
  pandas' `dt.year` returns int32, causing an int32-vs-int64 index
  mismatch in `pd.testing.assert_series_equal`. Production Parquet now
  agrees with its declared schema.

Test counts: tests/test_schemas.py 5 → 10 (5 raw-CSV + 5 Parquet-parametrised);
tests/test_aggregates.py 2 → 5 (2 raw + 3 Parquet row-conservation);
full suite 46 → 54 passing (+8; zero regression). All three derived-layer
test classes (schemas, aggregates, determinism) now green for their
Parquet variants.

+++ src/uk_subsidy_tracker/schemes/cfd/aggregation.py
@@ -64,7 +64,8 @@ def build_annual_summary(output_dir: Path) -> pd.DataFrame:
     has already collapsed to monthly.
     """
     sm = _read_station_month(output_dir)
-    sm["year"] = sm["month_end"].dt.year
+    # Cast to int64 to match AnnualSummaryRow.year declared dtype (D-10).
+    sm["year"] = sm["month_end"].dt.year.astype("int64")
 
     annual_cfd = (
         sm.groupby("year", sort=True)
@@ -89,7 +90,7 @@ def build_annual_summary(output_dir: Path) -> pd.DataFrame:
     daily = daily.dropna(subset=["counterfactual_total"])
     daily["cf_gbp"] = daily["counterfactual_total"] * daily["gen_mwh"]
     # daily_gen was grouped on Settlement_Date; the index name survives concat.
-    daily["year"] = daily.index.year
+    daily["year"] = daily.index.year.astype("int64")
     annual_cf = (
         daily.groupby("year", sort=True)["cf_gbp"]
         .sum()
@@ -120,7 +121,8 @@ def build_annual_summary(output_dir: Path) -> pd.DataFrame:
 def build_by_technology(output_dir: Path) -> pd.DataFrame:
     """Build and persist `by_technology.parquet` (year × technology grain)."""
     sm = _read_station_month(output_dir)
-    sm["year"] = sm["month_end"].dt.year
+    # Cast to int64 to match ByTechnologyRow.year declared dtype (D-10).
+    sm["year"] = sm["month_end"].dt.year.astype("int64")
 
     df = (
         sm.groupby(["year", "technology"], sort=True, dropna=False)
@@ -152,7 +154,8 @@ def build_by_allocation_round(output_dir: Path) -> pd.DataFrame:
     not carry it. Determinism preserved: identical raw → identical sum.
     """
     sm = _read_station_month(output_dir)
-    sm["year"] = sm["month_end"].dt.year
+    # Cast to int64 to match ByAllocationRoundRow.year declared dtype (D-10).
+    sm["year"] = sm["month_end"].dt.year.astype("int64")
 
     cfd_roll = (
         sm.groupby(["year", "allocation_round"], sort=True, dropna=False)
@@ -169,7 +172,7 @@ def build_by_allocation_round(output_dir: Path) -> pd.DataFrame:
     gen = load_lccc_dataset("Actual CfD Generation and avoided GHG emissions")
     gen = gen.dropna(subset=["CFD_Generation_MWh", "Strike_Price_GBP_Per_MWh"])
     gen = gen[gen["CFD_Generation_MWh"] > 0]
-    gen["year"] = gen["Settlement_Date"].dt.year
+    gen["year"] = gen["Settlement_Date"].dt.year.astype("int64")
     co2 = (
         gen.groupby(["year", "Allocation_round"], sort=True, dropna=False)[
             "Avoided_GHG_tonnes_CO2e"
